+++ changelog.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented to this file.
 Change Log history will be recorded in descending order
 
+## 0.0.3
+** Changed Files **
+- Added Home Controller
+- Added Stock Model
+- Loaded Stocks Database for the Homepage
 
 ## 0.0.2
 ** Changed Files **